requirements.txt
absl-py==0.9.0
alembic==1.3.2
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
matplotlib==3.1.2
mtcnn==0.1.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
Theano==1.0.4
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
